changed schema, and map js to add lat/lng to db

+++ db/migrate/20150330202012_add_lat_lng_to_spaces.rb
@@ -1,6 +1,6 @@
 class AddLatLngToSpaces < ActiveRecord::Migration
   def change
-    add_column :spaces, :lat, :integer
-    add_column :spaces, :lng, :integer
+    add_column :spaces, :lat, :float
+    add_column :spaces, :lng, :float
   end
 end
